+++ src/data/usecases/load-account-by-token/db-load-account-by-token-protocols.ts
@@ -1,0 +1,5 @@
+/* istanbul ignore file */
+export * from '@/data/protocols/db/account/load-account-by-token-repository'
+export * from '@/domain/models/account'
+export * from '@/data/protocols/criptography/decrypter'
+export * from '@/domain/usecases/load-account-by-token'
