+++ src/data/usecases/load-account-by-token/db-load-account-by-token.spec.ts
@@ -1,7 +1,5 @@
-import { LoadAccountByTokenRepository } from '@/data/protocols/db/account/load-account-by-token-repository'
-import { AccountModel } from '@/domain/models/account'
+import { LoadAccountByTokenRepository, AccountModel, Decrypter } from './db-load-account-by-token-protocols'
 import { DbLoadAccountByToken } from './db-load-account-by-token'
-import { Decrypter } from '@/data/protocols/criptography/decrypter'
 
 const makeDecrypter = (): Decrypter => {
   class DecrypterStub implements Decrypter {
