+++ src/data/usecases/load-survey-by-id/db-load-survey-by-id.spec.ts
@@ -1,7 +1,6 @@
-import { SurveyModel } from '@/domain/models/survey'
-import { DbLoadSurveyById } from './db-load-survey-by-id'
 import MockDate from 'mockdate'
-import { LoadSurveyByIdRepository } from '@/data/protocols/db/survey/load-survey-by-id-repository'
+import { DbLoadSurveyById } from './db-load-survey-by-id'
+import { SurveyModel, LoadSurveyByIdRepository } from './db-load-survey-by-id-protocols'
 
 const makeLoadSurveyByIdRepository = (): LoadSurveyByIdRepository => {
   class LoadSurveyByIdRepositoryStub implements LoadSurveyByIdRepository {
