+++ src/data/usecases/load-surveys/db-load-surveys-protocols.ts
@@ -1,0 +1,4 @@
+/* istanbul ignore file */
+export * from '@/data/protocols/db/survey/load-surveys-repository'
+export * from '@/domain/models/survey'
+export * from '@/domain/usecases/load-surveys'
